v4

+++ Models/Dto/ApplicationDto.cs
@@ -7,7 +7,7 @@
 		public int IdCourse { get; set; }
 		public int IdImageCourse { get; set; }
 		public int IdType { get; set; }
-		public int ScoreCourse { get; set; }
+		public double ScoreCourse { get; set; }
 		public bool LikeCourse { get; set; }
 		public string TitleCourse { get; set; } = string.Empty;
 		public string TypeName { get; set; } = string.Empty;
